Organized Files and added thread barriers
